requirements.txt
PyPDF2==3.0.1
requests==2.31.0
pinecone-client==3.2.0
python-dotenv==1.0.0
